Extract picker option button
